[todoApp]jsを整える
