+++ 03_todoApp/index.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
   <link rel="stylesheet" href="./src/css/style.css">
-  <script src="./src/javascript/script.js" type="module"></script>
+  <script src="./src/javascript/main.js" type="module"></script>
   <title>todoApp</title>
 </head>
 <body>
