+++ 03_todoApp/src/javascript/modules/displayNone.js
@@ -1,0 +1,3 @@
+export const displayNone = (item) => {
+  item.style.display = 'none';
+}
