--- 03_todoApp/src/javascript/toggleDisplay.js
@@ -1,12 +1,0 @@
-export const toggleDisplay = (state) => {
-  document.querySelectorAll('.finished').forEach((item) => {
-    switch (state) {
-      case 'hide':
-        item.classList.add('hide');
-        break;
-      case 'show':
-        item.classList.remove('hide');
-        break;
-    }
-  })
-};
